% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapPaths.R
\name{heatmapPaths}
\alias{heatmapPaths}
\title{Visualization of the gene set enrichment meta-analysis results}
\usage{
heatmapPaths(
  objectMApath,
  resMA,
  scaling = c("zscor", "rscale", "swr", "none"),
  regulation = c("all", "up", "down"),
  breaks = c(-2, 2),
  fdrSig = 0.05,
  comES_Sig = 0.5,
  numSig = "all",
  color = colorRampPalette(rev(brewer.pal(n = 7, name = "RdYlBu")))(100),
  na_col = "darkgrey",
  legend = TRUE,
  cluster_cols = FALSE,
  cluster_rows = FALSE,
  show_rownames = TRUE,
  show_colnames = FALSE,
  fontsize = 10,
  fontsize_row = fontsize,
  fontsize_col = fontsize
)
}
\arguments{
\item{objectMApath}{A list of list. Each list contains two elements.
The first element is the Gene Set matrix (gene sets in rows and samples in
columns) and the second element is a vector of zeros and ones that represents
the state of the different samples of the Gene Sets matrix.
0 represents one group (controls) and 1 represents the other group (cases).}

\item{resMA}{Output generated by the  function that performs
meta-analysis (metaAnalysisESpath).}

\item{scaling}{Character variable to choose between different scaling
approaches. See "Details" for more information.}

\item{regulation}{Character variable that indicates whether we want the
heatmap to show all significant paths ("all"), only the up-regulated paths
("up") or only the down-regulated paths("down")}

\item{breaks}{Numeric vector of length 2 that contains the extreme values
(minimum and maximum) of the range of values in which the heatmap
color scale will be distributed. Default a vector By default a
vector of -2 and 2 as extreme values.}

\item{fdrSig}{Adjusted p-value from which a gene set is considered
significant. Default 0.05}

\item{comES_Sig}{In absolute value. Combine effect size
threshold from which gene sets are considered. Default 0.5}

\item{numSig}{The number of most significant paths to be represented.
If numSig = "all", all significant paths that meet the selected parameters
will be represented.}

\item{color}{vector of colors used in heatmap.}

\item{na_col}{color of the NA cell in the heatmap.}

\item{legend}{logical to determine if legend should be drawn or not.}

\item{cluster_cols}{boolean values determining if columns should be
clustered.}

\item{cluster_rows}{boolean values determining if rows should be clustered.}

\item{show_rownames}{boolean specifying if row names are be shown.}

\item{show_colnames}{boolean specifying if column names are be shown.}

\item{fontsize}{base fontsize for the plot}

\item{fontsize_row}{fontsize for rownames (Default: fontsize)}

\item{fontsize_col}{fontsize for colnames (Default: fontsize)}
}
\value{
The matrix represented in the heatmap
}
\description{
It allows to see how the different significant gene sets are expressed in the
different samples
}
\details{
Scaling approaches that can be used are:
\itemize{
\item "rscale": it applies rescale function of \emph{scales} package. Values
will be between -1 and 1)
\item "zscor":  It calculates a z-score value for each gene, that is, the
mean gene expression from each gene is subtracted from each gene expression
value and then it is divided by the standard deviation
\item "swr": it applys scaling relative to reference dataset approach
\item "none": any scaling approach it is applied.
}
}
\examples{

data("simulatedData")
resultsMA <- metaAnalysisESpath(objectMApathSim, typeMethod="REM")
heatmapPaths(objectMA=objectMApathSim, resultsMA,
    scaling = "zscor", regulation = "all",breaks=c(-2,2),
    fdrSig = 0.05, comES_Sig = 1.5, numSig=20)

}
\references{
Hadley Wickham and Dana Seidel (2020). scales: Scale Functions for
Visualization. R package version 1.1.1.
\url{https://CRAN.R-project.org/package=scales}

Lazar, C, Meganck, S, Taminau, J, and et al. 2013. “Batch Effect
Removal Methods for Microarray Gene Expression Data Integration:
A Survey,” 469–90.

Raivo Kolde 2019. pheatmap: Pretty Heatmaps. R package version 1.0.12.
\url{https://CRAN.R-project.org/package=pheatmap}
}
\seealso{
\code{\link{createObjectMApath}}, \code{\link{metaAnalysisESpath}}
}
\author{
Juan Antonio Villatoro Garcia,
\email{juanantoniovillatorogarcia@gmail.com}
}
